requirements.txt
langchain
ipykernel
langchain_community
langchain_core
python-dotenv
pypdf
bs4
arxiv
pymupdf
langchain-text-splitters
sentence_transformers
langchain-huggingface
faiss-cpu
langchain_chroma
langchain_groq
fastapi
langserve
sse_starlette
pydantic==1.10.13
wikipedia
langgraph
streamlit
duckduckgo-search
psycopg2-binary
validators==0.28.1
youtube_transcript_api
pytube
unstructured
numexpr
langchain-astradb>=0.3.3
cassio
datasets
PyPDF2